regression test accept

+++ src/Repositories/ProjectRepository.php
@@ -75,7 +75,9 @@ class ProjectRepository
         $allMemberships = $member->tribeMemberships;
 
         foreach ($allMemberships as $membership) {
-            $membership->delete(); // todo select proiject
+            if ($membership->project_id === $project->id) {
+                $membership->delete();
+            }
         }
     }
